requirements.txt
aiofiles==24.1.0
aiogram==3.20.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.3
blinker==1.9.0
blur==0.4
celery==5.5.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.1.3
django-filter==24.1
django-phonenumber-field==8.3.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.11
Flask==3.1.2
frozenlist==1.5.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.4.3
packaging==25.0
pillow==11.0.0
prompt_toolkit==3.0.52
propcache==0.3.1
psycopg2-binary==2.9.11
psycopg2-pool==1.2
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.4
pytz==2025.2
PyYAML==6.0.3
redis==5.0.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
yarl==1.19.0
